requirements.txt
python-dateutil==2.8.2
Django==4.1.5
psycopg2==2.9.5
Pillow==9.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.0.0
